base64_encode {openssl}	R Documentation

Encode and decode base64

Description

Encode and decode binary data into a base64 string. Character vectors are automatically collapsed into a single string.

Usage

base64_encode(bin, linebreaks = FALSE)

base64_decode(text)

Arguments

bin
raw or character vector with data to encode into base64

linebreaks
insert linebreaks in the base64 message to make it more readable

text
string with base64 data to decode

Examples

input <- charToRaw("foo = bar + 5")
message <- base64_encode(input)
output <- base64_decode(message)
identical(output, input)

[Package openssl version 1.4.6 Index]