openssl_config {openssl}	R Documentation

OpenSSL Configuration Info

Description

Shows libssl version and configuration information.

Usage

openssl_config()

fips_mode()

Details

Note that the "fips" flag in openssl_config means that FIPS is supported, but it does not mean that it is currently enforced. If supported, it can be enabled in several ways, such as a kernel option, or setting an environment variable OPENSSL_FORCE_FIPS_MODE=1. The fips_mode() function shows if FIPS is currently enforced.

[Package openssl version 2.1.1 Index]